BackEnd/requirements.txt
asgiref==3.4.1
astroid==2.4.2
autopep8==1.5.7
colorama==0.4.4
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
importlib-metadata==4.6.1
isort==5.7.0
lazy-object-proxy==1.4.3
Markdown==3.3.4
mccabe==0.6.1
pycodestyle==2.7.0
pylint==2.6.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.10.0.0
wget==3.2
wrapt==1.12.1
zipp==3.5.0
